Fixing NPM build errors
